requirements.txt
Django==1.11
djangorestframework
django-filter
markdown
MarkupSafe
